+++ src/main/java/com/thorinhood/dbtg/repositories/SolutionsRepository.java
@@ -13,6 +13,6 @@ public interface SolutionsRepository extends JpaRepository<Solution, Object>, Jp
 
     @Query("SELECT r FROM com.thorinhood.dbtg.models.Solution r " +
         "WHERE ((:student is null) or (:student = student)) and ((:task is null) or (:task = task))")
-    List<Solution> studentOrTask(@Param("student") Long student, @Param("task") Integer task);
+    List<Solution> studentOrTask(@Param("student") Long student, @Param("task") Long task);
 
 }
